doDelete 에 권한설정

+++ src/main/java/com/sbs/exam/demo/Controller/UsrArticleController.java
@@ -71,13 +71,26 @@ public class UsrArticleController {
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession httpSession, int id) {
+		
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+
+		if (httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId =(int) httpSession.getAttribute("loginedMemberId");
+		}
+		
 		Article article = articleService.getArticle(id);
 
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 게시물은 존재하지 않습니다.", id));
 		}
-
+		
+		if(article.getMemberId() != loginedMemberId ) {
+			return ResultData.from("F-2", "해당글에 권한이 없습니다.");
+		}
+		
 		articleService.deleteArticle(id);
 
 		return ResultData.from("S-1", Ut.f("%d번 게시물을 삭제했습니다.", id));
